Separate out comparison tests
